refactor starcoder
