+++ server/text_generation_server/models/__init__.py
@@ -15,6 +15,7 @@ from text_generation_server.models.model import Model
 from text_generation_server.models.causal_lm import CausalLM
 from text_generation_server.models.bloom import BLOOM
 from text_generation_server.models.santacoder import SantaCoder
+from text_generation_server.models.starcoder import STARCODER
 
 from optimum.habana.transformers.modeling_utils import adapt_transformers_to_gaudi
 
@@ -86,13 +87,7 @@ def get_model(
     model_type = config_dict["model_type"]
 
     if model_type == "gpt_bigcode":
-        return SantaCoder(
-            model_id,
-            revision,
-            use_medusa=use_medusa,
-            dtype=dtype,
-            trust_remote_code=trust_remote_code,
-        )
+        return STARCODER(model_id, revision, dtype)
 
     if model_type == "bloom":
         return BLOOM(
